# apps/telomi/server/evaluation/operations-openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Telomi Operations HTTP",
		"version": "3.0.0",
		"description": "Loopback-only Evaluation Interface. Generated from server/evaluation/operations-contract.ts; do not edit by hand.",
		"x-protocol-version": 3,
		"x-schema-hash": "eb21cdf5e3dc53341cdb3655669adcbbfe3a47caa9c45ea1c7cecaac428cd5fb"
	},
	"servers": [
		{
			"url": "http://127.0.0.1:8788",
			"description": "Operations Listener (loopback only)"
		}
	],
	"paths": {
		"/operations/v1/agents": {
			"get": {
				"operationId": "getAgentCatalog",
				"summary": "Read product Agent names, source Bundles and output presentation contracts",
				"tags": [
					"capture"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AgentCatalogResponse"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/status": {
			"get": {
				"operationId": "getStatus",
				"summary": "Protocol handshake and Replay queue status",
				"tags": [
					"capture"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OperationsStatus"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/goals/{goalId}/cases": {
			"get": {
				"operationId": "listCases",
				"summary": "List captured Cases for a Goal",
				"tags": [
					"capture"
				],
				"parameters": [
					{
						"name": "goalId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "agentId",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CaseListResponse"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/goals/{goalId}/cases/{sourceRunId}/{caseId}": {
			"get": {
				"operationId": "readCase",
				"summary": "Read one Case manifest",
				"tags": [
					"capture"
				],
				"parameters": [
					{
						"name": "goalId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "sourceRunId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "caseId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CaseResponse"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/goals/{goalId}/cases/{sourceRunId}/{caseId}/files": {
			"get": {
				"operationId": "listCaseFiles",
				"summary": "List the Evidence files of one Case",
				"tags": [
					"capture"
				],
				"parameters": [
					{
						"name": "goalId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "sourceRunId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "caseId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CaseFileListResponse"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/goals/{goalId}/cases/{sourceRunId}/{caseId}/file": {
			"get": {
				"operationId": "readCaseFile",
				"summary": "Read one Evidence file of a Case by ref",
				"tags": [
					"capture"
				],
				"parameters": [
					{
						"name": "goalId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "sourceRunId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "caseId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "ref",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "File stream",
						"content": {
							"application/octet-stream": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/goals/{goalId}/cases/{sourceRunId}/{caseId}/bundle": {
			"get": {
				"operationId": "exportCaseBundle",
				"summary": "Export a Case Bundle tar",
				"tags": [
					"capture"
				],
				"parameters": [
					{
						"name": "goalId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "sourceRunId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "caseId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "File stream",
						"content": {
							"application/octet-stream": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/goals/{goalId}/cases/{sourceRunId}/{caseId}/provider-children/{executionId}/bundle": {
			"get": {
				"operationId": "exportProviderChildCaseBundle",
				"summary": "Export one Provider Child as a self-contained Case Bundle",
				"tags": [
					"capture"
				],
				"parameters": [
					{
						"name": "goalId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "sourceRunId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "caseId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "executionId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "File stream",
						"content": {
							"application/octet-stream": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/bundles/import": {
			"post": {
				"operationId": "importBundle",
				"summary": "Import a Case Bundle tar from the configured Exchange Root",
				"tags": [
					"eval-instance"
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BundleImportRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BundleImportResponse"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/goals/{goalId}/capability-snapshots/{snapshotId}": {
			"get": {
				"operationId": "readCapabilitySnapshot",
				"summary": "Read a Capability Snapshot manifest",
				"tags": [
					"eval-instance"
				],
				"parameters": [
					{
						"name": "goalId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "snapshotId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CapabilitySnapshotResponse"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/goals/{goalId}/replays": {
			"post": {
				"operationId": "startReplay",
				"summary": "Queue a Candidate Replay",
				"tags": [
					"eval-instance"
				],
				"parameters": [
					{
						"name": "goalId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReplayRequest"
							}
						}
					}
				},
				"responses": {
					"202": {
						"description": "Accepted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReplayRunResponse"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/goals/{goalId}/replays/{runId}": {
			"get": {
				"operationId": "readReplay",
				"summary": "Read a Replay Run",
				"tags": [
					"eval-instance"
				],
				"parameters": [
					{
						"name": "goalId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "runId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReplayRunResponse"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/goals/{goalId}/replays/{runId}/files": {
			"get": {
				"operationId": "readReplayFile",
				"summary": "Read one Trace or Artifact file of a Replay Run",
				"tags": [
					"eval-instance"
				],
				"parameters": [
					{
						"name": "goalId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "runId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "ref",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "File stream",
						"content": {
							"application/octet-stream": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/goals/{goalId}/replays/{runId}/evaluation-batch": {
			"get": {
				"operationId": "readEvaluationBatch",
				"summary": "Read the blind A/B batch of a Replay Run",
				"tags": [
					"eval-instance"
				],
				"parameters": [
					{
						"name": "goalId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "runId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EvaluationBatchResponse"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/goals/{goalId}/replays/{runId}/evaluation-batch/{pairId}/outputs/{label}/artifact": {
			"get": {
				"operationId": "readEvaluationOutputArtifact",
				"summary": "Read one blind A/B output artifact",
				"tags": [
					"eval-instance"
				],
				"parameters": [
					{
						"name": "goalId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "runId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "pairId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "label",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "file",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "File stream",
						"content": {
							"application/octet-stream": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/goals/{goalId}/replays/{runId}/executions/{executionId}/artifact": {
			"get": {
				"operationId": "readExecutionArtifact",
				"summary": "Read one execution artifact of a Replay Run",
				"tags": [
					"eval-instance"
				],
				"parameters": [
					{
						"name": "goalId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "runId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "executionId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "file",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "File stream",
						"content": {
							"application/octet-stream": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/operations/v1/goals/{goalId}/replays/{runId}/cancel": {
			"post": {
				"operationId": "cancelReplay",
				"summary": "Cancel a Replay Run",
				"tags": [
					"eval-instance"
				],
				"parameters": [
					{
						"name": "goalId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "runId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReplayRunResponse"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
...
			"AgentCatalogResponse": {
				"additionalProperties": false,
				"type": "object",
				"required": [
					"ok",
					"agents"
				],
				"properties": {
					"ok": {
						"const": true,
						"type": "boolean"
					},
					"agents": {
						"type": "array",
						"items": {
							"additionalProperties": false,
							"type": "object",
							"required": [
								"id",
								"displayName",
								"sourcePath",
								"presentationKind"
							],
							"properties": {
								"id": {
									"minLength": 1,
									"type": "string"
								},
								"displayName": {
									"minLength": 1,
									"type": "string"
								},
								"sourcePath": {
									"anyOf": [
										{
											"minLength": 1,
											"type": "string"
										},
										{
											"type": "null"
										}
									]
								},
								"presentationKind": {
									"anyOf": [
										{
											"const": "source-search",
											"type": "string"
										},
										{
											"const": "provider-child",
											"type": "string"
										},
										{
											"const": "note",
											"type": "string"
										},
										{
											"const": "investigation",
											"type": "string"
										},
										{
											"const": "report",
											"type": "string"
										},
										{
											"const": "wiki",
											"type": "string"
										},
										{
											"const": "evolution",
											"type": "string"
										},
										{
											"const": "chat",
											"type": "string"
										},
										{
											"const": "generic",
											"type": "string"
										}
									]
								},
								"impactPaths": {
									"type": "array",
									"items": {
										"minLength": 1,
										"type": "string"
									}
								},
								"replayPromptModes": {
									"type": "array",
									"items": {
										"anyOf": [
											{
												"const": "candidate",
												"type": "string"
											},
											{
												"const": "observed",
												"type": "string"
											},
											{
												"const": "override",
												"type": "string"
											}
										]
									}
								},
								"promptOverrideFields": {
									"type": "array",
									"items": {
										"anyOf": [
											{
												"const": "systemPrompt",
												"type": "string"
											},
											{
												"const": "userPrompt",
												"type": "string"
											}
										]
									}
								}
							}
						}
					}
				}
			},
...
			"OperationsStatus": {
				"additionalProperties": false,
				"type": "object",
				"required": [
					"ok",
					"protocolVersion",
					"schemaHash",
					"mode",
					"writable",
					"active",
					"queued",
					"concurrency",
					"recipes",
					"runtimeBuild",
					"runtimeBuildMatchesDisk",
					"agentBundleSha256",
					"capture"
				],
				"properties": {
					"ok": {
						"const": true,
						"type": "boolean"
					},
					"protocolVersion": {
						"minimum": 1,
						"type": "integer"
					},
					"schemaHash": {
						"pattern": "^[0-9a-f]{64}$",
						"type": "string"
					},
					"mode": {
						"anyOf": [
							{
								"const": "capture",
								"type": "string"
							},
							{
								"const": "eval",
								"type": "string"
							}
						]
					},
					"writable": {
						"type": "boolean"
					},
					"active": {
						"minimum": 0,
						"type": "integer"
					},
					"queued": {
						"minimum": 0,
						"type": "integer"
					},
					"concurrency": {
						"minimum": 1,
						"type": "integer"
					},
					"recipes": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"runtimeBuild": {
						"type": "string"
					},
					"runtimeBuildMatchesDisk": {
						"type": "boolean"
					},
					"agentBundleSha256": {
						"type": "string"
					},
					"capture": {
						"additionalProperties": false,
						"type": "object",
						"required": [
							"enabled",
							"failures",
							"recent",
							"retention"
						],
						"properties": {
							"enabled": {
								"type": "boolean"
							},
							"failures": {
								"minimum": 0,
								"type": "integer"
							},
							"recent": {
								"type": "array",
								"items": {
									"additionalProperties": false,
									"type": "object",
									"required": [
										"node",
										"reason",
										"at"
									],
									"properties": {
										"node": {
											"minLength": 1,
											"type": "string"
										},
										"reason": {
											"type": "string"
										},
										"at": {
											"minLength": 1,
											"type": "string"
										}
									}
								}
							},
							"retention": {
								"additionalProperties": false,
								"type": "object",
								"required": [
									"enabled",
									"maxAgeDays",
									"maxBytes",
									"cases",
									"bytes",
									"deletedByAge",
									"deletedBySize",
									"protectedActive",
									"protectedExporting",
									"warnings"
								],
								"properties": {
									"enabled": {
										"type": "boolean"
									},
									"maxAgeDays": {
										"minimum": 0,
										"type": "number"
									},
									"maxBytes": {
										"minimum": 0,
										"type": "number"
									},
									"lastSweepAt": {
										"minLength": 1,
										"type": "string"
									},
									"cases": {
										"minimum": 0,
										"type": "integer"
									},
									"bytes": {
										"minimum": 0,
										"type": "number"
									},
									"deletedByAge": {
										"minimum": 0,
										"type": "integer"
									},
									"deletedBySize": {
										"minimum": 0,
										"type": "integer"
									},
									"protectedActive": {
										"minimum": 0,
										"type": "integer"
									},
									"protectedExporting": {
										"minimum": 0,
										"type": "integer"
									},
									"warnings": {
										"type": "array",
										"items": {
											"type": "string"
										}
									}
								}
							}
						}
					}
				}
			},
...
			"BundleImportRequest": {
				"additionalProperties": false,
				"type": "object",
				"required": [
					"path"
				],
				"properties": {
					"path": {
						"minLength": 1,
						"type": "string"
					}
				}
			},
...
			"CaseFileListResponse": {
				"additionalProperties": false,
				"type": "object",
				"required": [
					"ok",
					"files"
				],
				"properties": {
					"ok": {
						"const": true,
						"type": "boolean"
					},
					"files": {
						"type": "array",
						"items": {
							"additionalProperties": false,
							"type": "object",
							"required": [
								"ref",
								"kind",
								"sha256",
								"byteLength"
							],
							"properties": {
								"ref": {
									"minLength": 1,
									"type": "string"
								},
								"kind": {
									"minLength": 1,
									"type": "string"
								},
								"sha256": {
									"pattern": "^[0-9a-f]{64}$",
									"type": "string"
								},
								"byteLength": {
									"minimum": 0,
									"type": "integer"
								}
							}
						}
					}
				}
			},
...
			"ErrorResponse": {
				"additionalProperties": true,
				"type": "object",
				"required": [
					"error"
				],
				"properties": {
					"error": {
						"type": "string"
					}
				}
			}
		}
	}
}